WIP: draft solution for putting action kwargs in function api and task
I would like to drop "abaqus_options" for "options", but this would break
backward compatibility and the tutorials. Adding "options" with a default
"${abaqus_options}" might be an ok transitionary change, but I think doing this
makes the default string substitution mandatory, as in, all task definitions
would be required to define "abaqus_options" in the task kwargs.

While making as few breaking changes as possible, keep "abaqus_options" for now
and try to work everything else into the function API.